requirements.txt
datumaro==0.1.11
Pillow>=8.3.2
python-dateutil==2.8.2
pytz==2021.1